[ide-service] fix verison resolver tests
